+++ src/rdma/common_client.cpp
@@ -213,7 +213,7 @@ int receive_server_metadata() {
 
   DEBUG("Server's swap area metadata received. addr: %p, len: %s, stag: 0x%x",
         (void *)swap_area_metadata.address,
-        human_readable_bytes((unsigned int)swap_area_metadata.length).c_str(),
+        human_readable_bytes(swap_area_metadata.length).c_str(),
         swap_area_metadata.stag.local_stag);
 
   return 0;
